Major refactorings and structure changes
