--- ax/Sources/AXHelper/AXUtils.swift
@@ -1,13 +1,0 @@
-// AXUtils.swift - Contains utility functions for accessibility interactions
-
-import Foundation
-import ApplicationServices
-import AppKit // For NSRunningApplication, NSWorkspace
-import CoreGraphics // For CGPoint, CGSize etc.
-
-// Constants like kAXWindowsAttribute are assumed to be globally available from AXConstants.swift
-// debug() is assumed to be globally available from AXLogging.swift
-// axValue<T>() is now in AXValueHelpers.swift
-
-// The file should be empty now except for comments and imports if all functions were moved.
-// If getElementAttributes and other core AX interaction functions are still here, they will remain.
