+++ ax/Sources/AXHelper/Utils/AXStringExtensions.swift
@@ -19,4 +19,13 @@ extension String {
 		}
 		return line
 	}
+}
+
+extension Optional {
+	var orNilString: String {
+		switch self {
+		case .some(let value): return "\(value)"
+		case .none:            return "nil"
+		}
+	}
 }
